refactor: MVC 패턴 분리
