+++ src/controller/AgeSimulation.java
@@ -1,0 +1,7 @@
+package controller;
+public class AgeSimulation {
+    public static void main(String[] args) {
+        AgeCalculatorController controller = new AgeCalculatorController();
+        controller.run();
+    }
+}
